Moved things in , cleanup from code review
