+++ lib/activity.js
@@ -10,7 +10,7 @@ var _ = require("highland"),
     AWS = require("aws-sdk");
 
 var activityHandler = require("./activity-handler"),
-    ActivityWorker = require("./lib/activity-worker");
+    ActivityWorker = require("./activity-worker");
 
 var agent = new https.Agent({
   // Infinity just boosts the max value; in practice this will be no larger
@@ -38,7 +38,7 @@ var swf = new AWS.SWF();
 module.exports = function(options, fn) {
   assert.ok(options.domain, "options.domain is required");
 
-  if(!fn) {
+  if (!fn) {
     assert.ok(options.activitiesFolder, "options.activitiesFolder required if no handler specified");
     fn = activityHandler(options.activitiesFolder, options.workerId);
   }
