+++ lib/decider.js
@@ -9,8 +9,8 @@ var assert = require("assert"),
 var _ = require("highland"),
     AWS = require("aws-sdk");
 
-var DeciderWorker = require("./lib/decider-worker"),
-    SyncDeciderWorker = require("./lib/sync-decider-worker");
+var DeciderWorker = require("./decider-worker"),
+    SyncDeciderWorker = require("./sync-decider-worker");
 
 AWS.config.update({
   region: process.env.AWS_DEFAULT_REGION || AWS.config.region || "us-east-1"
@@ -87,6 +87,7 @@ module.exports = function(options, fn) {
     // TODO get events from an LRU cache (which means requesting events in
     // reverse order)
     // in batches of 100 (or whatever the page size is set to)
+    console.log("POLLING.. %s %s", process.env.AWS_DEFAULT_REGION, options.domain);
     var poll = swf.pollForDecisionTask({
       domain: options.domain,
       taskList: {
@@ -94,6 +95,7 @@ module.exports = function(options, fn) {
       },
       identity: util.format("swfr@%s:%d", os.hostname(), process.pid)
     }, function(err, data) {
+      console.log(data);
       if (err) {
         console.warn(err.stack);
