+++ lib/shell.js
@@ -1,6 +1,7 @@
 "use strict";
 
-var execFile = require("child_process").execFile;
+var execFile = require("child_process").execFile,
+    os = require("os");
 
 var debug = require("debug"),
     holdtime = require("holdtime");
@@ -34,7 +35,8 @@ module.exports = function(cmd, args, options, callback) {
         }
       };
 
-  process.setMaxListeners(0);
+  // Only complain if we're trying to start more shell jobs than CPUs.
+  process.setMaxListeners(os.cpus().length * 2);
 
   // terminate child processes when the runner is asked to quit; successfully
   // completed processes will carry on and upload their output
