+++ lib/sync-decider-context.js
@@ -33,7 +33,7 @@ SyncDeciderContext.prototype.activity = function() {
         return reject(new Error(util.format("Unable to locate activity: %s@%s", name, version)));
       }
 
-      return activity.apply(context, args.concat(function(err, result) {
+      return activity.apply(null, args.concat(function(err, result) {
         if (err) {
           return reject(err);
         }
